% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.pstest.R
\name{print.pstest}
\alias{print.pstest}
\title{Print}
\usage{
\method{print}{pstest}(x, ...)
}
\arguments{
\item{x}{A pstest object}

\item{...}{Other params (required as generic function, but not used)}
}
\description{
Prints a pstest Object
}

