% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstest.package.R
\docType{package}
\name{pstest-package}
\alias{pstest-package}
\title{pstest: An R Package for Assessing the Goodness-of-fit of Propensity Score Models.}
\description{
An R package for implementing all the tests proposed in Sant'Anna and
Song (2016), 'Specification Tests for the Propensity Score', available at
Pedro H.C. Sant'Anna webpage, \url{http://sites.google.com/site/pedrohcsantanna/} .
}
\details{
In short, this package implement bootstrapped based Kolmogorov-Smirnov
and Cramer-von Mises type tests for parametric propensity score models with
either logistic ('logit'), or standard normal ('probit') link function.

The module was written by Pedro H.C. Sant'Anna (Vanderbilt University).

A help file with examples will be made available in the near future.
}

