% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstest.R
\name{pstest}
\alias{pstest}
\title{pstest: Specification Tests for the Propensity Score}
\usage{
pstest(d, pscore, xpscore, model = "logit", nboot = 1000, cores = 1)
}
\arguments{
\item{d}{vector containing the binary treatment indicator}

\item{pscore}{vector containing the fitted propensity score}

\item{xpscore}{matrix (or data frame) containing the covariates (and their
transformations) included in the propensity score
estimation. You should always include a constant term.}

\item{model}{functional form (link function) employed while estimating the propensity score.
The alternatives are: 'logit' (default), and 'probit'.}

\item{nboot}{number of bootstrap draws. Default is 1,000.}

\item{cores}{number of cores to use during the bootstrap (default is 1).
If cores>1, the bootstrap is conducted using parLapply, instead
of lapply type call.}
}
\value{
a list containing the Kolmogorov-Smirnov and Cramer-von Mises test
       statistics for the null hypothesis of correctly specified propensity
       score model (kstest and cvmtest, respectively), and their associate bootstrapped
       p-values, pvks and pvcvm, respectively.
}
\description{
\emph{pstest} computes Kolmogorov-Smirnov and Cramer-von Mises type tests
for the null hypothesis that a parametric model for the propensity score is
is correctly specified. For details of the testing procedure, see
Sant'Anna and Song (2016),'Specification Tests for the Propensity Score'.
}

